pyvector code cleaning
